refactor: normalize names of pages, add barrel file for pages and components, remove default export of App | bh
