+++ src/pages/CollectionPage.tsx
@@ -13,11 +13,9 @@ import {
     PSpinner
 } from "@porsche-design-system/components-react";
 
-export function ComponentsCollection() {
-
+export const CollectionPage = (): JSX.Element => {
     const [submit, setSubmit] = useState(false);
     const [activePage, setActivePage] = useState(1);
-
 
     const handleSubmit = (e: React.MouseEvent<HTMLPButtonElement>): void => {
         e.preventDefault();
