+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export * from './CollectionPage';
+export * from './FormsPage';
+export * from './PhnHeaderPage';
