+++ src/tests/App.test.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { fireEvent, render } from '@testing-library/react';
-import { ComponentsCollection } from "../pages/ComponentsCollection";
-import { Forms } from "../pages/Forms";
+import { CollectionPage, FormsPage } from "../pages";
 
 test('renders a headline from Porsche Design System', async () => {
-    const {getByText} = render(<ComponentsCollection/>);
+    const {getByText} = render(<CollectionPage />);
     const submitButtonElement = getByText("Submit");
     fireEvent.click(submitButtonElement);
     const headLineElement = getByText(/Hello/i);
@@ -12,7 +11,7 @@ test('renders a headline from Porsche Design System', async () => {
 });
 
 test('dismisses the headline from Porsche Design System', async () => {
-    const {getByText} = render(<ComponentsCollection/>);
+    const {getByText} = render(<CollectionPage />);
     const submitButtonElement = getByText("Submit");
     fireEvent.click(submitButtonElement);
     const headLineElement = getByText("Hello");
@@ -23,7 +22,7 @@ test('dismisses the headline from Porsche Design System', async () => {
 });
 
 test('headline should be changed according the selected value', async () => {
-    const {getByText, getByTestId} = render(<Forms/>);
+    const {getByText, getByTestId} = render(<FormsPage />);
     expect(getByText("Change this Headline by selecting")).toBeInTheDocument();
 
     fireEvent.change(getByTestId('select'), {target: {value: 'Headline B'}});
@@ -34,7 +33,7 @@ test('headline should be changed according the selected value', async () => {
 });
 
 test('headline should be displayed after click on Checkbox', async () => {
-    const {getByText, getByTestId} = render(<Forms/>);
+    const {getByText, getByTestId} = render(<FormsPage />);
     const input = getByTestId('checkbox');
 
     input.click();
@@ -42,7 +41,7 @@ test('headline should be displayed after click on Checkbox', async () => {
 });
 
 test('headline should be displayed after click on RadioButton', async () => {
-    const {getByText, getByTestId} = render(<Forms/>);
+    const {getByText, getByTestId} = render(<FormsPage />);
     const input = getByTestId('radiobutton');
 
     input.click();
@@ -50,7 +49,7 @@ test('headline should be displayed after click on RadioButton', async () => {
 });
 
 test('headline should be changed according the typed value', async () => {
-    const {getByText, getByTestId} = render(<Forms/>);
+    const {getByText, getByTestId} = render(<FormsPage />);
     getByText("Change this Headline by typing");
 
     fireEvent.change(getByTestId("input"), {target: {value: 'Headline C'}});
@@ -62,27 +61,27 @@ test('headline should be changed according the typed value', async () => {
 
 /* jsdom has some limitations. One of them is the fact that we cant change location. But we are able to test the closest href */
 test('slotted Link should navigate to PDS while mocked', async () => {
-    const {getByText} = render(<Forms/>);
+    const {getByText} = render(<FormsPage />);
 
     expect(getByText('Slotted Link').closest('a')).toHaveAttribute('href', 'https://designsystem.porsche.com')
 });
 
 test('a wrapped Link should navigate to #hashTest', async () => {
-    const {getByText} = render(<ComponentsCollection/>);
+    const {getByText} = render(<CollectionPage />);
     const link = getByText(/Test PLinkPure/i);
 
     expect(link.closest('a')).toHaveAttribute('href', '#hashTest')
 });
 
 test('a wrapped Link should navigate to #hashTest', async () => {
-    const {getByText} = render(<ComponentsCollection/>);
+    const {getByText} = render(<CollectionPage />);
     const link = getByText(/Test propHash/i);
 
     expect(link.closest('a')).toHaveAttribute('href', '#propHashTest')
 });
 
 test('pagination should return page 2', async () => {
-    const {container, getByText} = render(<ComponentsCollection/>);
+    const {container, getByText} = render(<CollectionPage />);
 
     if (!container.querySelector('li[value=NEXT_PAGE_LINK]')){
         return;
