+++ src/tests/SingleComponent.test.tsx
@@ -1,14 +1,13 @@
 import { render } from '@testing-library/react';
 import React from 'react';
-import { SingleComponent } from '../pages/SingleComponent';
-
+import { SingleComponent } from '../components';
 
 jest.mock('@porsche-design-system/components-react', () => {
     return require('@porsche-design-system/components-react/mocks/p-headline-mock');
 });
 
 test('renders a headline from Porsche Design System', async () => {
-    const {getByText} = render(<SingleComponent/>);
+    const {getByText} = render(<SingleComponent />);
     const headLineElement = getByText('Show single mock usage');
     expect(headLineElement).toBeInTheDocument();
 });
